+++ changes.md
@@ -1,5 +1,10 @@
 # Morbid
 
+ - Emitting tokens for service accounts
+ - Scala upgrade 3.3.3 -> 3.7.2
+ - Configurable days for tokens
+ - Added `POST /emit`
+
 ## Release v1.5.0
 LTS: 05/10/2025